Replaced the routing management code.
Rather than applying the whole router at once, it's now applied on a per-channel basis, this allows for buttons (such as the fader mutes) to make small changes without needing to reapply the full router.

In addition, after reading the routing from the config and before applying it follows a path which allows other settings to make 'transient' modifications..

Finally, putting this all together has introduced the 'Mute to X' behaviour on the mute buttons allowing for all 3 states (Unmuted, Mute to X, Mute All) to be supported.