Big chunk of changes, fully implement Cough Button
This ended up being a little bigger than expected, per-windows behaviour the button needs to be isolated from any faders, and due to the <MuteChat /> tag having different attributes to control the same behaviours quite a bit of code has had to be duplicated.

With that said, it should be fully functional with the 'Mute to X' transient routing behaviour working.