Removed Mute State from Mixer and CLI (see extended comment).
This is a little controversial, however there's no longer an 'easy' way to represent, or set, mute states in a way that can be stored and persisted in the profile for channels which are not attached to faders. For channels which *ARE* attached to faders, you can just push (or hold) the button on the GoXLR. Removing this was necessary to properly support and persist the three different mute 'states' represented by the GoXLR, as well as introduce behaviours and functionality which are consistent with expected behaviour based on how the device functions under Windows.