Initial commit, proper sampler handling.
Utilises goxlr-audio.sh to determine sinks / sources, as well as handling playback and recording. Currently supports pulseaudio (via pulse or pipewire-pulse), but could be expanded to support alsa, or even ffmpeg if you're on MacOS.

Currently only basic playback is supported, based on the samples configured in your profile. The files need to exist in your samples directory to work.

The goxlr-audio.sh script is currently expected to be in /usr/share/goxlr/ or ~/.local/share/goxlr-utility/ (this will eventually be automated).