Replaced initialisation error with a warning.
The error was used previously as 'out of the box' generally required a reboot after init. The error would appear on screen (placing the device into the 'ignore' list for a period of time) allowing users to note the problem, and act accordingly.

These days, we have better support through ALSA and UCM, so there's no real point in delaying the load, so instead drop a warning in case people are using JACK and proceed with the load.