Huge refactor of Profile Validation
Extended comment on this one, previously value validation was happening somewhere between device.rs and profile.rs in an attempt to ensure only correct values were passed to the GoXLR profile.

This change moves that validation into the profile crate itself. Not only does that make the profile crate more portable, but it also centralises validation to where it should really be.

In addition, when loading profiles, (where applicable) value validation now also happens, confirming the XML in the profile is valid.

So far, this works with all profiles I've sent at the GoXLR, although there may be some edge cases!