Resolved Pitch Encoder Issues
I tried to do this as cleanly as possible, but it's a little messy. In the profile, 'knob position' actually means *VALUE*, where the actual knob position varies depending on the 'mode'.

This code mostly moves the calculations across to the profile, with the daemon only setting up and rechecking when appropriate.