Finalised temporary code for devices that don't report SN
This is going to need to be revisited. When more than one device is plugged in that doesn't report a serial number, we have no real way of differentiating them. We can potentially look into using the USB port as a faux serial number in that scenario, although it isn't ideal.

Fortunately, this is going to be an *EXTREME* edge case, the probability of someone having a device without a serial is rare, the probability of them having TWO is extremely unlikely! So this can be looked into at a later date.

Shout out to 'Avid (ebuttonsdude)#2121' on Discord for the help with this!