Updated Audio Handler to be more permissive.
Comment on this one, the GoXLR Audio Handler previously flat out failed if it couldn't find an output device. This makes sense until you factor in Pulse / Pipewire, which may take a few seconds to actually create the sinks after login.

This change means that so long as the script exists, any interaction with the sampler after a failure will cause a re-check so long as 5 seconds have passed. This will hopefully catch 99% of cases where sink creation is delayed.