Update the Effect order to represent Windows behaviour, send different FX types as different commands.
Additional comment for this one, during testing of FX Preset importing (and profile set loading), I was noticing that sometimes things sounded a little weird, and loading appeared to be inconsistent.
 On closer review on the Windows side, there's a very specific execution order of the FX on the GoXLR (for example, in almost all cases a 'Style' parameter will appear before others of the same type.

This change primarily replicates that order, and to ensure the order, the HashSet is replaced with a LinkedHashSet for setting Effect Keys.