When initialising, remember to read the buffer if device already initialised.
When running multiple devices, if the initialiser is being used to initialise, it can place one of the devices into an 'unusable' state by not correctly reading the response of the test command.

This is only really an issue if the Daemon is actively running while initialisation occurs (for example, under the current MacOS project).