Final Cleanup of Initialisation Issue
This goes hand in hand with f164b3ff, if the initialiser is run while the daemon is active, the command index will be reset. This code will attempt to correct and resync the command index with the daemon to save needing to 'disconnect' and 'reconnect' the device to correct.