Fix Global and LogoX being the wrong way around.
From testing (now I can change these), the 'Global' colour is the colour of the LED inside the physical GoXLR unit. 'LogoX' (defined as 'Accent' in the UI) is simply the Colour of the X on the front of the GoXLR.

These weren't reversed particularly well (it was known they existed, but not the specifics), so this change corrects the settings.