Expose expected UI states to IPC
This is primarily important due to the compressor, when in 'Simple' mode, calculations are handled by the daemon based on the 'Amount' value, when in complex mode they're handled by the regular values.

While I'd prefer for clients to manage this themselves, these are the 'expected' display configurations based on the Windows App.