Force identifier requirement on Websocket Requests
Because the HTTP server contains multiple workers, and different commands can take different amounts of time to return valid data, it's very possible, and very easy, to accidentally make responses to request arrive out-of-order. As such, from the client side it's impossible to asynchronously support 'Do X, then Y with the result'.

This change forces clients to send an identifier with their request, which will be included in the response, to allow mappings.