Update Environment every 5 seconds under Linux...
This is a little ugly, xdg-open is used to launch things like web browsers and path openers, however it's very possible that the needed environmental variables may not have been set when launching the daemon on Login.

This code will periodically check the current state of the environment (via systemctl --user show-environment) and transpose the necessary values into the runtime environment if they're not already set.

In the event systemd is not available, this code will instead quietly fail.