Removed 'Pause' between file events
The pause was causing behaviours like mass delete events to not appropriately propagate up to the DaemonStatus, resulting in removed files being left behind in the list when the 'Update' happened in the middle of the delete process. Events are a little spammier now, but it shouldn't have a large effect on performance.