Slightly better handling of Default profiles
Now that we have the `goxlr-defaults` binary, Default.goxlr and DEFAULT.goxlrMicProfile are going to be present on the file system for the first load, we might as well use them.

The embedded versions of these are now only used as a fallback, in the event that either a profile is missing, or (more importantly) a profile cannot be loaded due to errors.