[EXPERIMENTAL] Maintain an open read to the Sampler
A lot of the problems related to the samplers 'Responsiveness' mostly are caused by the ability to create a new file, activate the WAV encoder, and (more importantly) actually open the Audio Stream from the OS.

That last step has been a little problematic for a while, as it can take up to half a second to detect and open the stream, by which point what you've wanted to record has been somewhat missed, this is unfortunately a bad experience.

This commit changes the sampler behaviour to be the same as the official app, when a device is connected we'll open it's sampler channel, and keep that channel open so that as soon as a record button is pressed we immediately start tracking samples. A buffer is in place to ensure that the period of time it takes to create the new file and establish the encoder is considered when writing the file.

This should provide a more optimal sampler experience.