Support a 'Pre-Button' buffer.
Initial code for this, not wired in yet. The util maintains an X (configurable) millisecond audio buffer which will be attached to the start of the recorded audio file when a sample record button is pressed.

This is useful for cases where something happens or is said, and you'd like to capture that!