Add a tiny pause between connecting to pulse and playing audio
So it seems calling Simple::new runs somewhat asynchronously, returning potentially before it's actually connected as a client. This can, in some cases, cause the first 50-100ms of the audio file to be lost on playback (although oddly, only the first time).

I've added a small pause into the connect state to hopefully try and mitigate this somewhat, I may need to move to full Pulse though.