Restore 'classic' settings when saving profile
The submix config format is only present if running the submix beta app, if you're monitoring, say, Stream Mix when saving a profile and attempt to load that into the official release version of the app, the Headphone routing will match Stream Mix.

This change ensures that when the config is saved, whatever you're monitoring gets reset to Headphones, and the routing table set appropriately. So then when loaded with something that doesn't support Monitoring, the routing and setup is correct.