Try harder to keep websocket alive on error.
This attempts primarily to extract the request `id` from the JSON in the event the Deserialise to Object fails. If it's able to do the parsing error will now get sent to the Websocket, rather than instantly closing the connection.

In addition, a CloseReason has been added to explain why a websocket has been forcefully closed.