Have settings save to a temporary file first.
As with profiles, this prevents edge cases where the util may be in the middle of writing the settings when the util is exited, causing the settings to be corrupt, and the util having to restore from defaults.

These are extreme edge cases, but they've occurred two or three times so far, so the goal is to help protect against them.