Turn settings paths into explicit default override
Backstory here, had a couple of cases where users have moved / copied their Utility settings to a new user, this leaves all the paths invalid as they'll be pointed to the old user.

The new code supports 'None' as a path option, if that's set it will assume a default directory under the current user, so moving things around should work cleanly unless the user has explicitly changed the value.

The code will also do a check on startup to see if the current set value is the same as the default path, if so, set the path to None and assume the defaults.