Better Sample Restart Handling (Thanks bryaudio!).
Originally Reported by SGTBubbaGump.

For context, when you spam the crap out of a button with a single Audio file attached, previously the original playback was completely torn down and the channel closed prior to reopening the same file, opening a new channel and beginning the read.

So instead of that overhead, when we determine we're restarting playback, we simply tell Symphonia to seek back to the beginning of the track and continue on.