Restart the Audio Handler every minute if there are no useful samples.
Windows will not enter Sleep mode if there is an application actively listening to an audio input, the best way to allow it to go to Sleep is to simply stop listening (even if it's for a fraction of a second).

This code will periodically check the defined sampler pre-buffer for *ANYTHING* that can be actively recorded, and if hitting a record button would result in a failure due to no usable samples being present, stop and immediately restart the listener.

With this behaviour, Windows will immediately go to Sleep as soon as the Sample listener is closed.

Shoutout to Paolo for suggesting the 'dead mans switch' :p