Improve audio recording loop to capture final audio chunk
The audio recording loop has been updated to better manage the ending part of the audio recording. This change includes the insertion of a loop to ensure that the last 'chunk' of audio, which may have been partially recorded when the stop button was released, is not lost. A conditional break is also added, taking into account the stop state of the recorder.